PDjango/requirements.txt
# Run: pip install -r requirements.txt

Django==2.2
djangorestframework==3.9.2
django-filter==2.1.0
